requirements.txt
Django==3.1.3
django-bootstrap4
django-ckeditor
Pillow
django-storages
boto3
pymysql
django-allauth
requests
coolsms_python_sdk
djangorestframework
iamport-rest-client